refactor(ocr): extract model provisioning into engine/models
